+++ docs/changelog.md
@@ -4,4 +4,12 @@ All notable changes to this project will be documented below.
 
 #### geospatial.read_geotif
 
-* v0.1dev: spectral = **geospatial.read_geotif**(*filename, bands="B,G,R"*)
+* v0.1dev: spectral = **geospatial.read_geotif**(*filename, bands="B,G,R", cropto=None*)
+
+#### geospatial.transform_points
+
+* v0.1dev: coord = **geospatial.transform_points**(*img, geojson*)
+
+#### geospatial.transform_polygons
+
+* v0.1dev: coord = **geospatial.transform_polygons**(*img, geojson*)